pyproject.toml
[project]
name = "tripletex-agent"
version = "0.1.0"
description = "AI agent for the NM i AI competition - interprets accounting prompts and executes Tripletex API calls"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "requests>=2.32.0",
    "openai>=1.50.0",
    "python-dotenv>=1.0.0",
    "pymupdf>=1.24.0",
    "pydantic>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "responses>=0.25.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "e2e: end-to-end tests requiring sandbox credentials",
]
